refactor: extract PlainWorkEntryRow, useExpandedFileDiff hook, unify adapter extraction
